Add a read-only Transifex service for development.
Now by default the Transifex service used in development is read-only,
trying to invoke the /push entry point will do nothing but write
an entry in the log. This can be disabled (so that data is actually pushed)
by setting the PUSH_TO_TRANSIFEX_ON_DEV configuration key to "False".